adopt std::bind_front in cases where it can simplify std::bind usage

+++ scrubby/gui/scrubbyeditor.h
@@ -40,7 +40,7 @@ public:
 	void outputChannelsChanged(size_t inChannelCount) override;
 
 private:
-	void HandleNotesButton(long inNotesButtonType);
+	void HandleNotesButton(size_t inNotesButtonType);
 	void HandlePitchConstraintChange();
 	void HandleTempoSyncChange();
 	void HandleTempoAutoChange();
